+++ docs/index-files/index-7.html
@@ -1163,7 +1163,7 @@
 <dt><span class="memberNameLink"><a href="../vn/loitp/restapi/uiza/model/v2/getdetailentity/GetDetailEntity.html#GetDetailEntity--">GetDetailEntity()</a></span> - Constructor for class vn.loitp.restapi.uiza.model.v2.getdetailentity.<a href="../vn/loitp/restapi/uiza/model/v2/getdetailentity/GetDetailEntity.html" title="class in vn.loitp.restapi.uiza.model.v2.getdetailentity">GetDetailEntity</a></dt>
 <dd>&nbsp;</dd>
     <dt><span class="memberNameLink"><a
-            href="../vn/loitp/restapi/uiza/util/UizaV3Util.html#getDetailEntity-vn.loitp.core.base.BaseActivity-java.lang.String-vn.loitp.restapi.uiza.util.UizaV3Util.Callback-">getDetailEntity(BaseActivity, String, UizaV3Util.Callback)</a></span>
+            href="../vn/loitp/restapi/uiza/util/UizaV3Util.html#getDetailEntity-vn.loitp.core.base.BaseActivity-java.lang.String-vn.loitp.uizavideov3.UizaV3Util.Callback-">getDetailEntity(BaseActivity, String, UizaV3Util.Callback)</a></span>
         - Static method in class vn.loitp.restapi.uiza.util.<a
                 href="../vn/loitp/restapi/uiza/util/UizaV3Util.html"
                 title="class in vn.loitp.restapi.uiza.util">UizaV3Util</a></dt>
