+++ docs/vn/loitp/restapi/uiza/util/UizaV3Util.html
@@ -107,7 +107,7 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>vn.loitp.restapi.uiza.util.UizaV3Util</li>
+                <li>vn.loitp.uizavideov3.UizaV3Util</li>
             </ul>
         </li>
     </ul>
@@ -200,7 +200,7 @@ extends java.lang.Object</pre>
                             <tr id="i1" class="rowColor">
                                 <td class="colFirst"><code>static void</code></td>
                                 <td class="colLast"><code><span class="memberNameLink"><a
-                                        href="../../../../../vn/loitp/restapi/uiza/util/UizaV3Util.html#getDetailEntity-vn.loitp.core.base.BaseActivity-java.lang.String-vn.loitp.restapi.uiza.util.UizaV3Util.Callback-">getDetailEntity</a></span>(<a
+                                        href="../../../../../vn/loitp/restapi/uiza/util/UizaV3Util.html#getDetailEntity-vn.loitp.core.base.BaseActivity-java.lang.String-vn.loitp.uizavideov3.UizaV3Util.Callback-">getDetailEntity</a></span>(<a
                                         href="../../../../../vn/loitp/core/base/BaseActivity.html"
                                         title="class in vn.loitp.core.base">BaseActivity</a>&nbsp;activity,
                                     java.lang.String&nbsp;entityId,
@@ -351,7 +351,7 @@ extends java.lang.Object</pre>
                                 <pre>public static&nbsp;java.lang.String&nbsp;getAppId(android.content.Context&nbsp;context)</pre>
                             </li>
                         </ul>
-                        <a name="getDetailEntity-vn.loitp.core.base.BaseActivity-java.lang.String-vn.loitp.restapi.uiza.util.UizaV3Util.Callback-">
+                        <a name="getDetailEntity-vn.loitp.core.base.BaseActivity-java.lang.String-vn.loitp.uizavideov3.UizaV3Util.Callback-">
                             <!--   -->
                         </a>
                         <ul class="blockListLast">
