+++ docs/vn/loitp/uizavideo/view/util/UizaUtil.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>vn.loitp.uizavideo.view.util.UizaUtil</li>
+<li>vn.loitp.uizavideov3.UizaUtil</li>
 </ul>
 </li>
 </ul>
