+++ uizacoresdk/src/main/java/vn/loitp/uizavideo/TrackSelectionHelper.java
@@ -45,10 +45,9 @@ import java.util.Arrays;
 
 import loitp.core.R;
 import vn.loitp.core.utilities.LAnimationUtil;
-import vn.loitp.core.utilities.LLog;
 import vn.loitp.core.utilities.LUIUtil;
 import vn.loitp.uizavideo.view.util.DemoUtil;
-import vn.loitp.uizavideo.view.util.UizaUtil;
+import vn.loitp.uizavideov3.UizaUtil;
 
 /**
  * Helper class for displaying track selection dialogs.
